requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
flake8==7.2.0
mccabe==0.7.0
pycodestyle==2.13.0
pyflakes==3.3.2
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.1
